Phase 2 Wave 4: split Rounds, RoundBudget, Dispatch, Submission, SubmissionQuery, BuildLog, BuildMedia slices out of store.ts
Completes the Store split (Phase 2 exit criteria): the last four
entangled interfaces sharing the submissions Map are now their own
slice files under store/slices/, with InMemoryStore/FirestoreStore
delegating every method to per-slice classes.
