+++ eslint-rules/module-size-baseline.json
@@ -401,7 +401,7 @@
     "apps/api/src/submission-token.test.ts": 24,
     "apps/api/src/submission-token.ts": 52,
     "apps/api/src/submissions.test.ts": 7172,
-    "apps/api/src/submissions.ts": 6116,
+    "apps/api/src/submissions.ts": 6118,
     "apps/api/src/suggestion-inbox.test.ts": 369,
     "apps/api/src/suggestion-inbox.ts": 325,
     "apps/api/src/suggestion-outcomes.test.ts": 205,
